backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
pandas==2.2.2
google-genai>=0.8.0
python-dotenv==1.0.1
httpx>=0.27.0
